refactor: create colorinit module
