+++ clovers/src/textures/spatial_checker.rs
@@ -6,7 +6,8 @@ use palette::convert::IntoColorUnclamped;
 use palette::white_point::E;
 use palette::Xyz;
 
-use super::{ColorInit, TextureTrait, TypedColorInit};
+use super::TextureTrait;
+use crate::colorinit::{ColorInit, TypedColorInit};
 use crate::{Float, Vec3, PI};
 
 /// A standard checkered texture based on spatial 3D texturing.
